# Bytecode generation for MapTransformValueFunction