This is related to https://github.com/prestodb/presto/issues/7436

Benchmark results:

Before:
```
Benchmark                                          (name)   (type)  Mode  Cnt   Score   Error  Units
BenchmarkTransformValue.benchmark        transform_values   BIGINT  avgt   20  64.830 ± 2.420  ns/op
BenchmarkTransformValue.benchmark        transform_values   DOUBLE  avgt   20  62.140 ± 2.225  ns/op
BenchmarkTransformValue.benchmark        transform_values  VARCHAR  avgt   20  89.144 ± 3.883  ns/op
BenchmarkTransformValue.benchmark  exact_transform_values   BIGINT  avgt   20  50.635 ± 2.510  ns/op
BenchmarkTransformValue.benchmark  exact_transform_values   DOUBLE  avgt   20  46.683 ± 0.580  ns/op
BenchmarkTransformValue.benchmark  exact_transform_values  VARCHAR  avgt   20  74.325 ± 1.145  ns/op
```



After:
```
Benchmark                                          (name)   (type)  Mode  Cnt   Score   Error  Units
BenchmarkTransformValue.benchmark        transform_values   BIGINT  avgt   20  55.218 ± 1.116  ns/op
BenchmarkTransformValue.benchmark        transform_values   DOUBLE  avgt   20  53.021 ± 2.259  ns/op
BenchmarkTransformValue.benchmark        transform_values  VARCHAR  avgt   20  81.336 ± 2.749  ns/op
BenchmarkTransformValue.benchmark  exact_transform_values   BIGINT  avgt   20  50.420 ± 1.304  ns/op
BenchmarkTransformValue.benchmark  exact_transform_values   DOUBLE  avgt   20  47.965 ± 1.345  ns/op
BenchmarkTransformValue.benchmark  exact_transform_values  VARCHAR  avgt   20  77.922 ± 1.746  ns/op
```