@wenleix 2017-05-08
Rebased and update to use the new MapType (similar to the change in https://github.com/wenleix/presto/commit/623eeb8561ac1a04f51539b4c265a4ef9f8883ab#diff-b56cfc7677611a8bb0dca1e8584041ec ). Will push it once the Jenkins is cleared.